added backend check for no active borrow request while deleting user

+++ backend/src/main/java/com/librarymanagementsys/backend/service/UserService.java
@@ -134,6 +134,11 @@ public class UserService {
         if (user == null) {
             throw new UserNotFoundException("User not found with email: " + email);
         }
+
+        if(user.getNoBooksBorrowed() > 0){
+            throw new IllegalStateException("Cannot delete user with borrowed books");
+        }
+
         try{
             userRepository.delete(user);
         } catch (Exception e) {
